+++ src/main/kotlin/io/github/jamalam/database/entity/AnnouncementSubscriber.kt
@@ -1,0 +1,7 @@
+package io.github.jamalam.database.entity
+
+import dev.kord.common.entity.Snowflake
+
+data class AnnouncementSubscriber(
+    val channel: Long
+)